Fix multiple async packages not working together
Future continuation callback is a single global callback on Rust side.
Its used to signal bindings from Rust when a future is ready. The
problem is if multiple async packages set this callback, only one of
these callbacks will be used by Rust code for futures from both
packages.

The current implementation of this callback simply sends the
continuation future result into a channel, but the channel type was
using `C.int8_t`, which turns out to be a different type for each
package, and channel raises an error when wrong type is being sent in
the channel.

Luckily this callback is very simple, and it doesn't matter which
package's callback is trigerred. So the fix is to use a channel type
that is the same for both packages, regular `int8` instead of `C.int8_t`.

+++ bindgen/templates/Async.go
@@ -3,8 +3,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */#}
 
 const (
-	uniffiRustFuturePollReady      C.int8_t = 0
-	uniffiRustFuturePollMaybeReady C.int8_t = 1
+	uniffiRustFuturePollReady      int8 = 0
+	uniffiRustFuturePollMaybeReady int8 = 1
 )
 
 func uniffiRustCallAsync(
@@ -111,8 +111,8 @@ func uniffiRustCallAsyncInner(
 	pollFunc func(*C.void, unsafe.Pointer, *C.RustCallStatus),
 	freeFunc func(*C.void, *C.RustCallStatus),
 ) (*C.void, error) {
-	pollResult := C.int8_t(-1)
-	waiter := make(chan C.int8_t, 1)
+	pollResult := int8(-1)
+	waiter := make(chan int8, 1)
 	chanHandle := cgo.NewHandle(waiter)
 
 	rustFuture, err := rustCallWithError(converter, func(status *C.RustCallStatus) *C.void {
@@ -146,8 +146,8 @@ func uniffiRustCallAsyncInner(
 //export uniffiFutureContinuationCallback{{ config.package_name.as_ref().unwrap() }}
 func uniffiFutureContinuationCallback{{ config.package_name.as_ref().unwrap() }}(ptr unsafe.Pointer, pollResult C.int8_t) {
 	doneHandle := *(*cgo.Handle)(ptr)
-	done := doneHandle.Value().((chan C.int8_t))
-	done <- pollResult
+	done := doneHandle.Value().((chan int8))
+	done <- int8(pollResult)
 }
 
 func uniffiInitContinuationCallback() {
